requirements.txt
Django==4.0.10
django-cors-headers
django-environ
django-filter
django-jazzmin
djangorestframework
djangorestframework-simplejwt
drf-spectacular==0.27.2
pillow
PyJWT
PyYAML
requests